save-config-pro_1.0-1/opt/save-config-pro/requirements.txt
paramiko
psutil
Pillow
matplotlib
